% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.add.rownames2dat}
\alias{bic.add.rownames2dat}
\title{Add a column to matrix containing its row names}
\usage{
bic.add.rownames2dat(dat, h = "")
}
\arguments{
\item{dat}{matrix}

\item{h}{optional header for new column}
}
\value{
matrix containing new extra column made of existing matrix's row names
}
\description{
Create a new column containing the row names of the matrix
and add it as the first column in the matrix. Takes in optional
second argument to set the name of the new column.
}
