% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.average.by.name}
\alias{bic.average.by.name}
\title{Average all values in a matrix that have the same 'name'}
\usage{
bic.average.by.name(dat)
}
\arguments{
\item{dat}{a matrix where the first column may contain duplicate values
(strings or numbers) and the second through Nth columns contain 
only numbers}
}
\value{
a matrix containing averages of values whose names were the
        the same in the original matrix, plus the original values in the 
        second through Nth columns of the input matrix whose 'names' were 
        unique
}
\description{
Given a two-column matrix where the first column may contain duplicate 
values ('names'), for every one of those duplicate values find the average 
of the corresponding numbers in the second column
}
