% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.check.alignment.distribution.data}
\alias{bic.check.alignment.distribution.data}
\title{Validate alignment distribution data from CollectRnaSeqMetrics}
\usage{
bic.check.alignment.distribution.data(dat)
}
\description{
Check that data frame contains the correct column names and that
all data is numeric
}
