% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.check.line.chart.data}
\alias{bic.check.line.chart.data}
\title{Check formatting of input to bic.plot.rseqc.line.chart()}
\usage{
bic.check.line.chart.data(dat, dat2 = NULL)
}
\description{
Ensure object(s) passed to 'dat' and/or 'dat2' arguments of bic.plot.rseqc.line.chart()
are data frames where column one contains integers in order from least to greatest,
and each remaining column contains metric values for one sample
}
