% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.check.picard.data}
\alias{bic.check.picard.data}
\title{Validate data from PICARD CollectRnaSeqMetrics or AlignmentSummaryMetrics}
\usage{
bic.check.picard.data(dat, name)
}
\arguments{
\item{dat}{data frame passed to plot function}

\item{name}{name of plot (Options: ["alignment.distribution" |
"5prime3prime.bias" |
"alignment.summary"]}
}
\description{
Check that data frame contains the correct column names and that
all data is numeric
}
