% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.check.read.distribution.data}
\alias{bic.check.read.distribution.data}
\title{Check formatting of input to bic.plot.read.distribution()}
\usage{
bic.check.read.distribution.data(dat)
}
\arguments{
\item{dat}{data frame passed to plot function}
}
\description{
Ensure object(s) passed to 'dat' argument of bic.plot.read.distribution()
is a data frame containing a 'Samples' slot and at least one other slot
containing data to plot
}
