% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R, R/tmp_bic_plots.R
\name{bic.deseq.heatmap}
\alias{bic.deseq.heatmap}
\alias{bic.deseq.heatmap}
\title{Draw a heatmap of counts table}
\usage{
bic.deseq.heatmap(cds, file = NULL, num.gns = 100, transform = FALSE)

bic.deseq.heatmap(cds, file = NULL, num.gns = 100, transform = FALSE)
}
\arguments{
\item{cds}{DESeq countsDataSet object}

\item{file}{PDF file to which heatmap should be saved (optional)}

\item{num.gns}{Include this number of most highly expressed genes in 
the heatmap; Default: 100}

\item{transform}{logical indicating whether to transform counts; 
Default: FALSE}

\item{cds}{DESeq countsDataSet object}

\item{file}{PDF file to which heatmap should be saved (optional)}

\item{num.gns}{Include this number of most highly expressed genes in 
the heatmap; Default: 100}

\item{transform}{logical indicating whether to transform counts; 
Default: FALSE}
}
\description{
Draw heatmap of either raw counts or variance stabilization
transformed (?) data for all samples. Write plot to PDF file.

Draw heatmap of either raw counts or variance stabilization
transformed (?) data for all samples. Write plot to PDF file.
}
