% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.deseq.plot.pca}
\alias{bic.deseq.plot.pca}
\title{Plot PCA}
\usage{
bic.deseq.plot.pca(cds, file = NULL)
}
\arguments{
\item{cds}{DESeq countDataSet}

\item{file}{PDF file to which plot should be saved}
}
\description{
Run DESeq \code{plotPCA()} on variance stabilised tranformed data
}
