% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.file2matrix}
\alias{bic.file2matrix}
\title{Read data file and return as a matrix}
\usage{
bic.file2matrix(dat, header = FALSE, sep = "", skip = 0)
}
\arguments{
\item{dat}{file to read}

\item{header}{logical to indicate whether file contains a header. Default: FALSE}

\item{sep}{delimiter in file}

\item{skip}{number of lines to skip before reading file}
}
\description{
Read data file and return as a matrix
}
