% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_analyze_counts.R
\name{bic.get.gene.sets}
\alias{bic.get.gene.sets}
\title{Get the gene sets that meet criteria set by cut offs}
\usage{
bic.get.gene.sets(gsaRes, gsa.tab, fc2keep = log2(1.5), frac2keep = 2,
  dir = "Up")
}
\arguments{
\item{gsaRes}{results from \code{piano::runGSA()}}

\item{gsa.tab}{BIC-ified GSA results}

\item{fc2keep}{minimum absolute value fold change; Default: log2(1.5)}

\item{frac2keep}{at least 1/frac2keep genes in set must have FC >= fc2keep}

\item{dir}{direction ["Up"|"Dn"] of the results in gsaRes}
}
\value{
filtered gsa.tab
}
\description{
Get the gene sets that meet criteria set by cut offs
}
