% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.get.package.info}
\alias{bic.get.package.info}
\title{Get info for one or multiple packages installed in current R version}
\usage{
bic.get.package.info(pkgs, fields = NULL)
}
\arguments{
\item{pkgs}{a vector of package names to look up}

\item{fields}{a vector of information fields to include in results; Default:
all available fields}
}
\value{
A data frame of all packages and fields for which information
        could be found, or NULL if no installed packages match
        package name(s)
}
\description{
Given a vector of package names and fields to look up, return a 
data frame with the requested information
}
