% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.grid.arrange.shared.legend}
\alias{bic.grid.arrange.shared.legend}
\title{Arrange multiple plots on one page with one shared legend}
\usage{
bic.grid.arrange.shared.legend(..., ncol = length(list(...)), nrow = 1,
  main = "", position = c("bottom", "right"))
}
\arguments{
\item{...}{ggplot objects}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{main}{main title}

\item{position}{position of legend ["bottom"|"right"]}
}
\description{
Place multiple plots on one page, with a shared legend either at the right or at the bottom
}
