% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.join.strings}
\alias{bic.join.strings}
\title{Convert a character vector to a delimited string}
\usage{
bic.join.strings(x, str)
}
\arguments{
\item{x}{character vector}

\item{str}{delimiter}
}
\value{
a string
}
\description{
Given a character vector and a delimiter, join all
items in the vector, separated by the given delimiter
}
