% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.make.colnames}
\alias{bic.make.colnames}
\title{Remove a row from matrix to use the values as column names}
\usage{
bic.make.colnames(dat, nn = 1)
}
\arguments{
\item{dat}{matrix containing extra row to be used as column names}

\item{nn}{optional row number to use as column names}
}
\value{
data matrix with new column names and one less row
}
\description{
Quick way to assign column headers to a matrix of data read
in from file. By default, the first row is used.
}
