% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.make.rownames}
\alias{bic.make.rownames}
\title{Remove a column from a matrix and use its values as the matrix's
rownames.}
\usage{
bic.make.rownames(dat, nn = 1)
}
\arguments{
\item{dat}{matrix containing extra column to be used as row names}

\item{nn}{column to be removed and used as row names}
}
\value{
data matrix with new rownames and one less column
}
\description{
Shortcut to assigning row names to a matrix of data read in from 
a file. By default, the first column of data is used.
}
