% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.matrix2numeric}
\alias{bic.matrix2numeric}
\title{Convert all values in matrix to numeric}
\usage{
bic.matrix2numeric(dat)
}
\arguments{
\item{dat}{matrix containing data to be converted}
}
\value{
numeric matrix
}
\description{
Quick way to ensure you are working with a numeric matrix
}
