% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.non.empty.file.found}
\alias{bic.non.empty.file.found}
\title{Search for a non-empty file by pattern}
\usage{
bic.non.empty.file.found(path, pattern)
}
\arguments{
\item{path}{full path to directory in which to search for file}

\item{pattern}{pattern to search}
}
\value{
NULL if file is not found or is empty, otherwise return 
        FIRST file name that matches pattern and is not empty. NOTE: if 
        multiple files match pattern, ONLY the FIRST file is returned,
        without warning
}
\description{
In the specified directory, search files for given pattern
and if found, determine if file is not empty
}
