% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R, R/tmp_bic_plots.R
\name{bic.pdf.hclust}
\alias{bic.pdf.hclust}
\alias{bic.pdf.hclust}
\title{Write PDF file containing heierarchical clustering tree}
\usage{
bic.pdf.hclust(dat, conds = NULL, file.name = NULL, title = "",
  width = 20, height = 14, lwd = 3, cex.main = 2.5, cex.lab = 3,
  cex = 3, xlab = "", ylab = "", sample.labels = NULL)

bic.pdf.hclust(dat, conds = NULL, file.name = NULL, title = "",
  width = 20, height = 14, lwd = 3, cex.main = 2.5, cex.lab = 3,
  cex = 3, xlab = "", ylab = "", sample.labels = NULL)
}
\arguments{
\item{dat}{matrix containing data to plot}

\item{conds}{vector of sample conditions, in same order as column names
in matrix or sample.labels; if given, nodes will be colored
according to conditions; Default: NULL}

\item{file.name}{file name. Default: "tmp.pdf"}

\item{title}{Title of plot}

\item{width}{plot width (see plot docs)}

\item{height}{plot height (see plot docs)}

\item{lwd}{line width}

\item{cex.main}{size of title}

\item{cex.lab}{size of labels}

\item{cex}{font size}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}

\item{sample.labels}{Vector of sample labels. Default: column names in matrix}

\item{dat}{matrix containing data to plot}

\item{file.name}{PDF file to which plot should be saved; Default: NULL}

\item{title}{Title of plot}

\item{sample.labels}{Vector of sample labels; Default: column names in matrix}

\item{conds}{vector of sample conditions, in same order as column names 
in matrix or sample.labels; if given, nodes will be colored
according to conditions; Default: NULL}

\item{width}{plot width (see plot docs)}

\item{height}{plot height (see plot docs)}

\item{lwd}{line width}

\item{cex.main}{size of title}

\item{cex.lab}{size of labels}

\item{cex}{font size}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}
}
\description{
Quickly plot heirarchical clustering of data with several default
parameters like width, height, font size, etc.

Quickly plot heirarchical clustering of data with several default
parameters like width, height, font size, etc.
}
