% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.plot.5prime3prime.bias}
\alias{bic.plot.5prime3prime.bias}
\title{Plot coverage uniformity}
\usage{
bic.plot.5prime3prime.bias(dat, pct = FALSE, stacked = TRUE,
  col.pal = "Set2", file = NULL)
}
\arguments{
\item{dat}{data frame consisting of data from CollectRNASeqMetrics
output}

\item{pct}{logical indicating that plot should show percentages}

\item{stacked}{logical indicating that bar chart should be stacked; Default: TRUE}

\item{col.pal}{name of color palette; must be from list of RColorBrewer palettes
Default: "Set2"}

\item{file}{PDF file to which plot should be saved (optional)}
}
\description{
Bar chart showing median CV coverage, median 5' and 3' bias,
and median 5' to 3' bias for each sample.
}
