% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R, R/tmp_bic_plots.R
\name{bic.plot.alignment.summary}
\alias{bic.plot.alignment.summary}
\alias{bic.plot.alignment.summary}
\title{Plot numbers of mapped and unmapped reads for each sample}
\usage{
bic.plot.alignment.summary(dat, position = "stack", pct = FALSE,
  col.pal = "Set3", file = NULL)

bic.plot.alignment.summary(dat, position = "stack", pct = FALSE,
  col.pal = "Set3", file = NULL)
}
\arguments{
\item{dat}{data frame of PICARD AlignmentSummaryMetrics, or at least 
containing columns CATEGORY, PF_READS, PF_READS_ALIGNED and 
SAMPLE, where CATEGORY contains at least values FIRST_OF_PAIR
and SECOND_OF_PAIR}

\item{position}{position of grouped bars; Default: "stack", option: "dodge"}

\item{pct}{logical indicating whether to show percentages rather than absolute
read counts}

\item{col.pal}{name of color palette; must be from list of RColorBrewer palettes
Default: "Set3"}

\item{file}{PDF file to which plot should be saved (optional)}

\item{dat}{data frame of PICARD AlignmentSummaryMetrics, or at least 
containing columns CATEGORY, PF_READS, PF_READS_ALIGNED and 
SAMPLE, where CATEGORY contains at least values FIRST_OF_PAIR
and SECOND_OF_PAIR}

\item{pct}{logical indicating whether to show percentages rather than absolute
read counts}

\item{col.pal}{name of color palette; must be from list of RColorBrewer palettes
Default: "Set3"}

\item{position}{position of grouped bars; Default: "stack", option: "dodge"}

\item{file}{PDF file to which plot should be saved (optional)}
}
\description{
Stacked bar chart showing either absolute values or percentages of mapped 
and unmapped reads for R1 and R2 of each sample. Takes in data frame
containing PICARD AlignmentSummaryMetrics, or at minimum a data frame with
columns: CATEGORY, PF_READS, PF_READS_ALIGNED, SAMPLE, where CATEGORY column
contains at least categories FIRST_OF_PAIR and SECOND_OF_PAIR.

Stacked bar chart showing either absolute values or percentages of mapped 
and unmapped reads for R1 and R2 of each sample. Takes in data frame
containing PICARD AlignmentSummaryMetrics, or at minimum a data frame with
columns: CATEGORY, PF_READS, PF_READS_ALIGNED, SAMPLE, where CATEGORY column
contains at least categories FIRST_OF_PAIR and SECOND_OF_PAIR.
}
