% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.plot.coverage}
\alias{bic.plot.coverage}
\title{Plot normalized coverage}
\usage{
bic.plot.coverage(dat, col.pal = "Spectral", file = NULL)
}
\arguments{
\item{dat}{data frame containing combined histograms from collectrnaseqmetrics}

\item{col.pal}{name of color palette; must be from list of RColorBrewer palettes
Default: "Spectral"}

\item{file}{PDF file to which plot should be saved (optional)}
}
\description{
Line chart showing normalized coverage for each sample
}
