% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R, R/tmp_bic_plots.R
\name{bic.plot.dispersion.estimates}
\alias{bic.plot.dispersion.estimates}
\alias{bic.plot.dispersion.estimates}
\title{Plot dispersion estimates}
\usage{
bic.plot.dispersion.estimates(cds, out.dir = getwd(), file.prefix = "")

bic.plot.dispersion.estimates(cds, out.dir = getwd(), file.prefix = "")
}
\arguments{
\item{cds}{DESeq countsDataSet}

\item{out.dir}{Directory in which plots should be saved; Default: $PWD}

\item{file.prefix}{String to prepend to PDF file names (optional)}

\item{cds}{DESeq countsDataSet}

\item{out.dir}{Directory in which plots should be saved; Default: $PWD}

\item{file.prefix}{String to prepend to PDF file names (optional)}
}
\description{
Plot dispersion estimates for all conditions in experiment. Draw one
plot per PDF file.

Plot dispersion estimates for all conditions in experiment. Draw one
plot per PDF file.
}
