% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.plot.ma}
\alias{bic.plot.ma}
\title{Plot MA}
\usage{
bic.plot.ma(res, file = NULL)
}
\arguments{
\item{res}{object returned from DESeq function nbinomTest}

\item{file}{PDF file to which plot should be saved (optional)}
}
\description{
Plot log2 fold changes versus mean of normalized counts for a DESeq
comparison
}
