% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.plot.rseqc.line.chart}
\alias{bic.plot.rseqc.line.chart}
\title{Plot RSeQC line metrics}
\usage{
bic.plot.rseqc.line.chart(dat, title, dat2 = NULL, title2 = NULL,
  main.title = NULL, col.pal = "Spectral", file = NULL)
}
\arguments{
\item{dat}{data frame where column1 contains read position for either 
read 1 or read 2, and the remaining columns are metric values
for each sample; position column name must be "Position"}

\item{title}{plot title}

\item{dat2}{second data frame with the same structure as dat1, to be plotted
adjacent to the plot of dat (optional; used to plot data for read 
1 and read 2 on one page)}

\item{title2}{title of second plot (optional)}

\item{main.title}{if plotting two datasets, this is the main title for the page}

\item{col.pal}{color palette (Default: "Spectral")}

\item{file}{PDF file to which plot should be saved (optional)}
}
\description{
Create a line chart for merged RSeQC metrics where the x axis is 
a numerical value and the y axis is a count
}
