% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_analyze_counts.R
\name{bic.process.gsa.res}
\alias{bic.process.gsa.res}
\title{Format and filter results from piano::runGSA()}
\usage{
bic.process.gsa.res(gsaRes, max.p = 0.1, fc2keep = log2(1.5),
  frac2keep = 2, fcQ = T)
}
\arguments{
\item{gsaRes}{result object from piano::runGSA()}

\item{max.p}{maximum p-value to use as cutoff; Default: 0.1}

\item{fc2keep}{minimum fold change}

\item{frac2keep}{at least 1/frac2keep genes in set must have FC >= fc2keep}

\item{fcQ}{remove gene sets that do not meet cutoffs set}
}
\value{
list of two matrices: one with up-regulated gene sets and one with down-regulated
        gene sets
}
\description{
Format and filter results from piano::runGSA()
}
