% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.pval.histogram}
\alias{bic.pval.histogram}
\title{Histogram of p-values}
\usage{
bic.pval.histogram(dat, file = NULL)
}
\arguments{
\item{dat}{data frame with at least one column with name 'pvals'}

\item{file}{PDF file to which plot should be saved (optional)}
}
\description{
Histogram of p-values
}
