% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.remove.sub}
\alias{bic.remove.sub}
\title{Remove a substring from a string}
\usage{
bic.remove.sub(x, substr = "hsa", part2take = 1)
}
\arguments{
\item{x}{character vector containing strings to 
be split}

\item{substr}{substring to be used as delimiter for
splitting}

\item{part2take}{if 1, return string preceding \code{substr}
if 2, return string trailing \code{substr}}
}
\value{
either the string preceding or trailing \code{substr}
}
\description{
For every string in vector x, split the string
on substring \code{substr} and return either
the string preceding trailing \code{substr}
}
