% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_analyze_counts.R
\name{bic.run.deseq.comparison}
\alias{bic.run.deseq.comparison}
\title{Run a differential expression comparison between two
conditions.}
\usage{
bic.run.deseq.comparison(countDataSet, conds, condA, condB, max.p = 0.05,
  min.abs.fc = 2, min.count = 10, zeroaddQ = F, genes = c())
}
\arguments{
\item{countDataSet}{A countDataSet object created by running DESeq's newCountDataSet()}

\item{conds}{Vector of sample conditons, in the same order as samples 
in raw count matrix}

\item{condA}{The first condition in comparison (must be in conditions vector)}

\item{condB}{The second condition in comparison (must be in conditions vector)}

\item{max.p}{A number, the max p-value cutoff to determine which genes are 
considered differentially expressed; Default: 0.05}

\item{min.abs.fc}{A number, the minimum absolute fold change cutoff to determine 
which genes are considered differentially expressed; Default: 2}

\item{min.count}{A number, minimum average number of reads in one condition; 
Default: 10}

\item{zeroaddQ}{Logical, indicates whether to include genes that have zero 
counts in one condition and insignificant p-value, but at 
least min.count/mean(sizeFactors) in the other condition; 
Default: FALSE}

\item{genes}{A named vector of gene symbols, where the names are equivalent 
to the IDs in the original raw count matrix (e.g., Ensembl IDs); 
Default: NULL}
}
\value{
List containing a matrix of ALL results, a matrix containing results
        for differentially expressed genes (as determined by cutoffs), a vector
        of DE genes, the cutoffs used in analysis, and norm factors (TO DO: reword this)
}
\description{
Compare expression in condition B vs expression in condition B. Results include
log2FC, adjusted p-val, and mean expression in each condition. Two result sets
are included in returned list: one including ALL genes, and one including only 
genes differentially expressed as determined by \code{max.p} and \code{min.abs.fc}. 
User has the option to include in the results those genes with zero counts in one
condition and count greater than \code{min.count/mean(sizeFactors)} in the other
condition, even if the p-value is greater than \code{max.p}, as those genes may still
be of interest.
}
