% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_analyze_counts.R
\name{bic.run.gsa}
\alias{bic.run.gsa}
\title{Run gene set analysis using the package 'piano'}
\usage{
bic.run.gsa(species, deseq.res, min.gns = 5, max.gns = 1000, max.p = 0.1,
  nPerm = 10000, fcQ = T, fc2keep = log2(1.5), frac2keep = 4)
}
\arguments{
\item{species}{currently only human and mouse are supported}

\item{deseq.res}{the matrix of ALL DESeq results (unfiltered)
in BIC format (GeneID,GeneSymbol,pval,P.adj,
log2[condB/condA],Mean_at_cond_condA,
Mean_at_cond_condB}

\item{min.gns}{minimum number of genes a set must have in 
order to be included in analysis; Default: 5}

\item{max.gns}{maximum number of genes a set must have in
order to be included in analysis; Default: 1000}

\item{max.p}{maximum p-value to use as cut off; Default: 0.1}

\item{nPerm}{the number of times the genes are randomized;
Default: 1e4}

\item{fcQ}{filter gene sets based on fc2keep and frac2keep
Default: TRUE}

\item{fc2keep}{fold change cutoff (see frac2keep); Default: 
log2(1.5)}

\item{frac2keep}{at least \code{1/frac2keep} of genes in a set
must have \code{fold change >= abs(fc2keep)};
Default: 4}
}
\value{
a list of two matrices: one with gene sets found to be
        significantly UP-regulated and one with gene sets found
        to be significantly DOWN-regulated
}
\description{
Run gene set analysis for one pairwise comparison of sample groups. 
GSA is run using the R package 'piano'. For each gene set, means 
are calculated for the genes with positive fold changes and for
genes with negative fold changes. P-values and adjusted p-values 
are then calculated and gene sets with a significant number of
of up- or down-regulated genes are reported in the output. Gene Sets
included are from the MSigDB Collections:
}
\details{
C1: positional gene sets
C2: curated gene sets
C3: motif gene sets
C5: GO gene sets
C6: oncogenic signatures
C7: immunologic signatures

Note: Mouse gene sets were downloaded from 
      http://bioinf.wehi.edu.au/software/MSigDB/
}
