% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.sample.to.sample.distances}
\alias{bic.sample.to.sample.distances}
\title{Draw a heatmap showing sample-to-sample distances using variance 
stabilized data}
\usage{
bic.sample.to.sample.distances(cds, conds, file = NULL)
}
\arguments{
\item{cds}{DESeq countDataSet}

\item{conds}{vector of sample conditions that was used to generate \code{cds}}

\item{file}{PDF file to which heatmap whould be saved (optional)}
}
\description{
Find potential sample mislabeling errors by viewing distances
between every pair of samples. Save plot to PDF file.
}
