% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_analyze_counts.R
\name{bic.scale.factor.quantile}
\alias{bic.scale.factor.quantile}
\title{Non-deseq way to normalize counts(?)}
\usage{
bic.scale.factor.quantile(raw.counts, percentile = "75\%")
}
\arguments{
\item{raw.counts}{Raw count matrix, where a column is one sample and a row is one gene}

\item{percentile}{A string representing percentile; e.g., "100\%". Default: "75\%"}
}
\value{
TO DO
}
\description{
TO DO: INSERT DESCRIPTION HERE
}
