% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R, R/tmp_bic_plots.R
\name{bic.standard.heatmap}
\alias{bic.standard.heatmap}
\alias{bic.standard.heatmap}
\title{Create PDF of generic red/green/black heatmap showing relative
expression values}
\usage{
bic.standard.heatmap(norm.counts.matrix, condA, condB, genes = NULL,
  file = NULL)

bic.standard.heatmap(norm.counts.matrix, condA, condB, genes = NULL,
  file = NULL)
}
\arguments{
\item{norm.counts.matrix}{matrix of normalized counts, where a row is a gene
and a column is a sample. May contain "GeneID" and/or
"GeneSymbol" column. If GeneSymbol column is present, 
it will be used for heatmap labeling. If not, GeneID 
column will be used.}

\item{condA}{the first condition named in the bic.deseq.file name}

\item{condB}{the second condition named in the bic.deseq.file name}

\item{genes}{a vector of genes to be included in the heatmap;
if normalized counts matrix includes "GeneSymbol" column,
genes must be from that column. Otherwise, they must
be from the "GeneID" column.}

\item{file}{name of PDF file to which heatmap should be written. (optional)}

\item{norm.counts.matrix}{matrix of normalized counts, where a row is a gene
and a column is a sample. May contain "GeneID" and/or
"GeneSymbol" column. If GeneSymbol column is present, 
it will be used for heatmap labeling. If not, GeneID 
column will be used.}

\item{condA}{the first condition named in the bic.deseq.file name}

\item{condB}{the second condition named in the bic.deseq.file name}

\item{genes}{a vector of genes to be included in the heatmap;
if normalized counts matrix includes "GeneSymbol" column,
genes must be from that column. Otherwise, they must
be from the "GeneID" column.}

\item{file}{name of PDF file to which heatmap should be written. (optional)}
}
\description{
Given the results of DESeq comparison of two conditions and the
full normalized count matrix, generate a heatmap showing the relative 
expression values of "top 100" differentially expressed genes. If
there are fewer than 100 genes in the results file, all of them will
be included in the heatmap. Color scheme is red/black/green.

Given the results of DESeq comparison of two conditions and the
full normalized count matrix, generate a heatmap showing the relative 
expression values of "top 100" differentially expressed genes. If
there are fewer than 100 genes in the results file, all of them will
be included in the heatmap. Color scheme is red/black/green.
}
