% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_util.R
\name{bic.write.dat}
\alias{bic.write.dat}
\title{Write file containing tab delimited data}
\usage{
bic.write.dat(dat, file.name, add.rownames = F, col.names = T, quote = F,
  sep = "\\t")
}
\arguments{
\item{dat}{matrix containing data to be writtedn}

\item{file.name}{file name}

\item{add.rownames}{include row names in file; Default: FALSE}

\item{col.names}{include column names in file; Default: TRUE}

\item{quote}{include quotes around each matrix cell; Default: FALSE}

\item{sep}{delimiter; Default: "\\t"}
}
\description{
Wrapper around \code{write.table()} with several defaults set.
}
