% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_analyze_counts.R
\name{bic.write.normalized.counts.file}
\alias{bic.write.normalized.counts.file}
\title{Write normalized counts matrix to file}
\usage{
bic.write.normalized.counts.file(norm.counts.mat, file.name)
}
\arguments{
\item{norm.counts.mat}{matrix containing normalized counts}

\item{file.name}{File to which results should be written}
}
\description{
Write normalized counts matrix to file
}
