% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R, R/tmp_bic_plots.R
\name{bic.check.line.chart.data}
\alias{bic.check.line.chart.data}
\alias{bic.check.line.chart.data}
\title{Check formatting of input to bic.plot.rseqc.line.chart()}
\usage{
bic.check.line.chart.data(dat, dat2 = NULL)

bic.check.line.chart.data(dat, dat2 = NULL)
}
\arguments{
\item{dat}{data frame passed to plot function}

\item{dat2}{second data frame passed to plot function (optional)}

\item{dat}{data frame passed to plot function}

\item{dat2}{second data frame passed to plot function (optional)}
}
\description{
Ensure object(s) passed to 'dat' and/or 'dat2' arguments of bic.plot.rseqc.line.chart()
are data frames where column one contains integers in order from least to greatest,
and each remaining column contains metric values for one sample

Ensure object(s) passed to 'dat' and/or 'dat2' arguments of bic.plot.rseqc.line.chart()
are data frames where column one contains integers in order from least to greatest,
and each remaining column contains metric values for one sample
}
