% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.hclust.samples}
\alias{bic.hclust.samples}
\title{Plot heirachical clustering of all samples}
\usage{
bic.hclust.samples(norm.counts, file.name = NULL)
}
\arguments{
\item{norm.counts}{data matrix containing normalized counts, where
a column represents a sample and a row represents
a gene. may or may not contain "GeneID" and "GeneSymbol"
columns.}

\item{file.name}{plot will be saved in this file; Default: 
$PWD/counts_scaled_hclust.pdf}
}
\description{
Plot clustering of samples using the heierarchical clustering method
}
