% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.pdf.hclust}
\alias{bic.pdf.hclust}
\title{Write PDF file containing heierarchical clustering tree}
\usage{
bic.pdf.hclust(dat, file.name = "tmp.pdf", title = "", width = 26,
  height = 16, lwd = 3, cex.main = 3, cex.lab = 3, cex = 3,
  xlab = "", ylab = "", sample.labels = NULL)
}
\arguments{
\item{dat}{matrix containing data to plot}

\item{file.name}{file name. Default: "tmp.pdf"}

\item{title}{Title of plot}

\item{width}{plot width (see plot docs)}

\item{height}{plot height (see plot docs)}

\item{lwd}{line width}

\item{cex.main}{size of title}

\item{cex.lab}{size of labels}

\item{cex}{font size}

\item{xlab}{label of x axis}

\item{ylab}{label of y axis}

\item{sample.labels}{Vector of sample labels. Default: column names in matrix}
}
\description{
Quickly plot heirarchical clustering of data with several default
parameters like width, height, font size, etc.
}
