% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.plot.5prime3prime.bias}
\alias{bic.plot.5prime3prime.bias}
\title{Plot coverage uniformity (5' and 3' bias)}
\usage{
bic.plot.5prime3prime.bias(dat, col.pal = "Set3", file = NULL)
}
\arguments{
\item{dat}{data frame consisting of data from CollectRNASeqMetrics
output}

\item{col.pal}{name of color palette; must be from list of RColorBrewer palettes
Default: "Set3"}

\item{file}{PDF file to which plot should be saved (optional)}
}
\description{
Bar chart showing median CV coverage, median 5' and 3' bias,
and median 5' to 3' bias for each sample.
}
