% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R
\name{bic.plot.alignment.distribution}
\alias{bic.plot.alignment.distribution}
\title{Plot distibution of alignment across different genomic locations}
\usage{
bic.plot.alignment.distribution(dat, pct = FALSE, horizontal = TRUE,
  col.pal = "Set3", file = NULL)
}
\arguments{
\item{dat}{data frame consisting of data from CollectRNASeqMetrics
output}

\item{pct}{plot percentages}

\item{horizontal}{logical indicating that bars should be horizontal; Default: TRUE}

\item{col.pal}{name of color palette; must be from list of RColorBrewer palettes
Default: "Set3"}

\item{file}{PDF file to which plot should be saved (optional)}
}
\description{
Bar chart showing for each sample the distribution of read
alignments across ribosomal, coding, UTR, intronic and 
intergenic bases
}
