% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_plots.R, R/tmp_bic_plots.R
\name{bic.plot.coverage}
\alias{bic.plot.coverage}
\alias{bic.plot.coverage}
\title{Plot normalized coverage}
\usage{
bic.plot.coverage(dat, col.pal = "Set3", file = NULL)

bic.plot.coverage(dat, col.pal = "Set3", file = NULL)
}
\arguments{
\item{dat}{data frame containing combined histograms from collectrnaseqmetrics}

\item{col.pal}{name of color palette; must be from list of RColorBrewer palettes
Default: "Set3"}

\item{file}{PDF file to which plot should be saved (optional)}

\item{dat}{data frame containing combined histograms from collectrnaseqmetrics}

\item{col.pal}{name of color palette; must be from list of RColorBrewer palettes
Default: "Set3"}

\item{file}{PDF file to which plot should be saved (optional)}
}
\description{
Line chart showing normalized coverage for each sample

Line chart showing normalized coverage for each sample
}
